# openapi.json
{"openapi":"3.0.1","info":{"title":"InWallet Portfolio API","description":"InWallet Finansal Portföy Yönetimi Mikroservis API Dokümantasyonu","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/transactions":{"post":{"tags":["transaction-controller"],"operationId":"createTransaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/goals":{"post":{"tags":["goal-controller"],"operationId":"createGoal","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Goal"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/budgets":{"post":{"tags":["budget-controller"],"operationId":"createBudget","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Budget"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Budget"}}}}}}},"/api/auth/verify":{"post":{"tags":["auth-controller"],"operationId":"verify","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logoutUser","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/assets":{"post":{"tags":["asset-controller"],"operationId":"createAsset","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Asset"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Asset"}}}}}}},"/api/transactions/user/{userId}":{"get":{"tags":["transaction-controller"],"operationId":"getTransactionsByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}},"/api/market/prices":{"get":{"tags":["market-controller"],"operationId":"getAllPrices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"number"}}}}}}}},"/api/goals/user/{userId}":{"get":{"tags":["goal-controller"],"operationId":"getGoalsByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Goal"}}}}}}}},"/api/budgets/user/{userId}":{"get":{"tags":["budget-controller"],"operationId":"getBudgets","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Budget"}}}}}}}},"/api/assets/user/{userId}":{"get":{"tags":["asset-controller"],"operationId":"getAssetsByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Asset"}}}}}}}},"/api/transactions/{id}":{"delete":{"tags":["transaction-controller"],"operationId":"deleteTransaction","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/goals/{id}":{"delete":{"tags":["goal-controller"],"operationId":"deleteGoal","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/budgets/{id}":{"delete":{"tags":["budget-controller"],"operationId":"deleteBudget","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/assets/{id}":{"delete":{"tags":["asset-controller"],"operationId":"deleteAsset","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Asset":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"type":{"type":"string"},"symbol":{"type":"string"},"quantity":{"type":"number"},"averageBuyPrice":{"type":"number"},"targetSellPrice":{"type":"number"},"currentPrice":{"type":"number"},"updatedAt":{"type":"string","format":"date-time"}}},"Goal":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"name":{"type":"string"},"initialPrice":{"type":"number"},"currentTargetPrice":{"type":"number"},"expectedInflationRate":{"type":"number"},"targetDate":{"type":"string","format":"date-time"},"completionPercentage":{"type":"number"},"targetAmount":{"type":"number"},"currentAmount":{"type":"number"}}},"Transaction":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"type":{"type":"string"},"description":{"type":"string"},"category":{"type":"string"},"asset":{"$ref":"#/components/schemas/Asset"},"amount":{"type":"number"},"pricePerUnit":{"type":"number"},"transactionDate":{"type":"string","format":"date-time"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"},"verificationCode":{"type":"string"},"monthlyIncome":{"type":"number"},"monthlyExpense":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"assets":{"type":"array","items":{"$ref":"#/components/schemas/Asset"}},"goals":{"type":"array","items":{"$ref":"#/components/schemas/Goal"}},"transactions":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}},"enabled":{"type":"boolean"}}},"Budget":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"category":{"type":"string"},"limitAmount":{"type":"number"},"period":{"type":"string"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}